requirements.txt
feedparser
python-dotenv
requests
openai
markdown
PyYAML
